Move port definitions to their own package
