+++ plugin/cmd/scheduler/scheduler.go
@@ -25,7 +25,7 @@ import (
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/client"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/client/record"
 	_ "github.com/GoogleCloudPlatform/kubernetes/pkg/healthz"
-	masterPkg "github.com/GoogleCloudPlatform/kubernetes/pkg/master"
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/master/ports"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/util"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/version/verflag"
 	"github.com/GoogleCloudPlatform/kubernetes/plugin/pkg/scheduler"
@@ -34,7 +34,7 @@ import (
 )
 
 var (
-	port         = flag.Int("port", masterPkg.SchedulerPort, "The port that the scheduler's http service runs on")
+	port         = flag.Int("port", ports.SchedulerPort, "The port that the scheduler's http service runs on")
 	address      = util.IP(net.ParseIP("127.0.0.1"))
 	clientConfig = &client.Config{}
 )
